moved common functions in utilities
