+++ config.ini
@@ -22,6 +22,7 @@
 'number_files' = 1
 #'schema' = 'schemas/basicHD5.json'
 #'verification_type' = 'hdf_tags'
+'dependencies' = 'schemas/dependencies.json'
 
 ['file_patterns']
 'extensions' = .txt, .hd5, .HD5, .hdf5, .HDF5, .h5, .H5
