+++ doc/source/api.rst
@@ -12,7 +12,7 @@ for `DQuality <https://github.com/bfrosik/data-quality>`_.
    api/dquality.pvverifier
    api/dquality.reporter
    api/dquality.fileverify
-   api/dquality.dependenciesverifier
+   api/dquality.dependency
 
 .. automodule:: dquality
    :members:
