--- doc/source/api/dquality.dependenciesverifier.rst
@@ -1,14 +1,0 @@
-:mod:`dquality.dependenciesverifier`
-====================================
-
-.. automodule:: dquality.dependenciesverifier
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-   .. rubric:: **Functions:**
-
-   .. autosummary::
-      find_value
-      verify_list
-      verify_dependencies
